resharper validations

+++ TenantSystem/BillPayment.cs
@@ -10,7 +10,7 @@ namespace TenantSystem
     {
         public int Id { get; set; }
         public int TenantId { get; set; }
-        public decimal AmountPaid { get; set; }
+        public double AmountPaid { get; set; }
         public DateTime PaymentDate { get; set; }
         public DateTime CreatedDate { get; set; }
         public string CreatedBy { get; set; }
